feat: unite unfavorite model
