+++ src/features/article/unfavorite-article/unfavorite-article.mutation.ts
@@ -12,7 +12,7 @@ export function useUnfavoriteArticleMutation(
     UseMutationOptions<
       Awaited<ReturnType<typeof FavoriteService.unfavoriteArticleMutation>>,
       DefaultError,
-      articleTypes.Article,
+      string,
       unknown
     >,
     'mutationKey' | 'onMutate' | 'onSuccess' | 'onError' | 'onSettled'
@@ -29,37 +29,74 @@ export function useUnfavoriteArticleMutation(
   return useMutation({
     mutationKey: ['article', 'unfavorite', ...mutationKey],
 
-    mutationFn: ({ slug }: articleTypes.Article) =>
-      FavoriteService.unfavoriteArticleMutation(slug),
+    mutationFn: (slug) => FavoriteService.unfavoriteArticleMutation(slug),
 
-    onMutate: async (updatedArticle) => {
+    onMutate: async (slug) => {
       await queryClient.cancelQueries({ queryKey: ArticleQueries.keys.root })
 
       const previousArticle = queryClient.getQueryData(
-        ArticleQueries.articleQuery(updatedArticle.slug).queryKey,
+        ArticleQueries.articleQuery(slug).queryKey,
       )
 
+      const updatedArticle = previousArticle && {
+        ...previousArticle,
+        favorited: false,
+        favoritesCount: previousArticle.favoritesCount - 1,
+      }
+
+      const previousInfiniteArticles =
+        queryClient.getQueriesData<articleTypes.InfiniteArticles>({
+          queryKey: ArticleQueries.keys.rootInfinity,
+        })
+
       queryClient.setQueryData(
-        ArticleQueries.articleQuery(updatedArticle.slug).queryKey,
+        ArticleQueries.articleQuery(slug).queryKey,
         updatedArticle,
       )
 
-      await onMutate?.(updatedArticle)
+      queryClient.setQueriesData(
+        { queryKey: ArticleQueries.keys.rootInfinity },
+        (infinityArticles: articleTypes.InfiniteArticles | undefined) => {
+          if (!infinityArticles) return
 
-      return { previousArticle }
+          const { pages, pageParams } = infinityArticles
+
+          const updatedPages = pages.map((articles) => {
+            if (!articles.has(slug)) return articles
+
+            const updatedArticles = new Map(articles)
+            const previousArticlePreview = articles.get(slug)!
+
+            const updatedArticlePreview = {
+              ...previousArticlePreview,
+              favorited: false,
+              favoritesCount: previousArticlePreview.favoritesCount - 1,
+            }
+
+            updatedArticles.set(slug, updatedArticlePreview)
+
+            return updatedArticles
+          })
+          return { pages: updatedPages, pageParams }
+        },
+      )
+
+      await onMutate?.(slug)
+
+      return { previousArticle, previousInfiniteArticles }
     },
 
     onSuccess,
 
-    onError: async (error, updatedArticle, context) => {
+    onError: async (error, slug, context) => {
       const { previousArticle } = context || {}
 
       queryClient.setQueryData(
-        ArticleQueries.articleQuery(updatedArticle.slug).queryKey,
+        ArticleQueries.articleQuery(slug).queryKey,
         previousArticle,
       )
 
-      await onError?.(error, updatedArticle, context)
+      await onError?.(error, slug, context)
     },
 
     onSettled: async (data, error, variables, context) => {
